+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 #### 2.0.9
 
+* [REST] added methods to validate JSON structure *2014-12-19*
+
+`seeResponseJsonMatchesJsonPath` validates response JSON against [JsonPath](http://goessner.net/articles/JsonPath/).
+Usage of JsonPath requires library "flow/jsonpath" to be installed.
+
+`seeResponseJsonMatchesXpath` validates response JSON against XPath.
+It converts JSON structure into valid XPath document and executes XPath for it.
+
+`grabDataFromResponseByJsonPath` method was added as well to grab data JSONPath.
+
 * Dependencies updated to Symfony 2.6.
 * [PhpBrowser][Frameworks] fixed `Unreachable field` error while filling [] fields in input and textarea fields. See #1585 *2014-12-18*